+++ src/lib/index.ts
@@ -2,3 +2,4 @@ export * from './module';
 
 export * from './dialog/index';
 export * from './snackbar/index';
+export * from './paginator/index';