+++ src/lib/paginator/paginator-intl.ts
@@ -1,0 +1,47 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {Injectable} from '@angular/core';
+import {Subject} from 'rxjs/Subject';
+
+/**
+ *
+ * Copy from angular material
+ *
+ * To modify the labels and text displayed, create a new instance of NwbPaginatorIntl and
+ * include it in a custom provider
+ */
+@Injectable()
+export class NwbPaginatorIntl {
+  /**
+   * Stream that emits whenever the labels here are changed. Use this to notify
+   * components if the labels have changed after initialization.
+   */
+  changes: Subject<void> = new Subject<void>();
+
+  /** A label for the page size selector. */
+  itemsPerPageLabel = 'Items per page:';
+
+  /** A label for the range of items within the current page and the length of the whole list. */
+  getRangeLabel = (page: number, pageSize: number, length: number) => {
+    if (length == 0 || pageSize == 0) {
+      return `0 of ${length}`;
+    }
+
+    length = Math.max(length, 0);
+
+    const startIndex = page * pageSize;
+
+    // If the start index exceeds the list length, do not try and fix the end index to the end.
+    const endIndex = startIndex < length ?
+      Math.min(startIndex + pageSize, length) :
+      startIndex + pageSize;
+
+    return `${startIndex + 1} - ${endIndex} of ${length}`;
+  }
+}